继续整理：未完成各类对象和内容元素的字符串化

+++ trunk/include/pdf/che_pdf_colorspace.h
@@ -27,7 +27,7 @@ public:
 	CHE_PDF_Color( CHE_Allocator * pAllocator = nullptr ) 
 		: CHE_Object( pAllocator ), mIndex(0)
 	{
-		for ( size_t i = 0; i < 4; ++i )
+		for ( uint32 i = 0; i < 4; ++i )
 		{
 			mComponent[i] = 0.0f;
 		}
@@ -62,12 +62,12 @@ public:
 		}
 	}
 
-	size_t GetComponentCount() const
+	uint32 GetComponentCount() const
 	{
 		return mIndex;
 	}
 
-	FLOAT GetComponent( size_t index ) const
+	FLOAT GetComponent( uint32 index ) const
 	{
 		if ( index < mIndex )
 		{
@@ -84,7 +84,7 @@ public:
 
 private:
 	FLOAT				mComponent[4];
-	size_t				mIndex;
+	uint32				mIndex;
 };
